% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Examplehighdimdiffindiff.R
\name{Examplehighdimdiffindiff}
\alias{Examplehighdimdiffindiff}
\title{An example to implement Dr-did estimator}
\usage{
Examplehighdimdiffindiff(p = 50, q = 8, n0 = 500, method = "Pol", rho.X = 0.5)
}
\arguments{
\item{p}{dimension of the linear specification}

\item{q}{parameter for the nonparametric specification}

\item{n0}{number of sample size}

\item{method}{basis function in nonparametric specification:
\itemize{
\item (default) Pol: Polynomial basis with parameter q/2 as the degree of polynomial;
\item Tri: Trigonometric polynomials with parameter q/2 as the degree of polynomial;
}}

\item{rho.X}{parameter to construct the correlation matrix for the high dimensional covariates}
}
\description{
An example to implement Dr-did estimator
}
