% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdimdiffindiff_crossfit.R
\name{highdimdiffindiff_crossfit}
\alias{highdimdiffindiff_crossfit}
\title{Perform doubly robust diff-in-diff estimator under high dimesional covariates}
\usage{
highdimdiffindiff_crossfit(y0, y1, treat, x, z, k = 3, method = "Pol", q)
}
\arguments{
\item{y0}{outcome at time 0}

\item{y1}{outcome at time 1}

\item{treat}{treatment status at time 1}

\item{x}{pretreatment high dimensional covariates}

\item{z}{pretreatment low diemsioanl covariates allowing nonparametric specification}

\item{k}{number of sample split: default is 3}

\item{method}{basis function in nonparametric specification:
\itemize{
\item (default) Pol: Polynomial basis with parameter q/2 as the degree of polynomial;
\item Tri: Trigonometric polynomials with parameter q/2 as the degree of polynomial;
}}

\item{q}{paramter for the basis function in nonparametric specification}
}
\value{
\item{xdebias}{debiased estimator for the high dimensional coefficients}
\item{gdebias}{debiased estimator for the nonparametric coefficients}
\item{stdx}{standard error for the high dimensional coefficients}
\item{stdg}{standard error for the nonparametric coefficients}
}
\description{
Perform doubly robust diff-in-diff estimator under high dimesional covariates
}
\author{
Sida Peng
}
