% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdimdiffindiff_crossfit_inside.R
\name{highdimdiffindiff_crossfit_inside}
\alias{highdimdiffindiff_crossfit_inside}
\title{The core function of doubly robust diff-in-diff estimator under high dimesional covariates}
\usage{
highdimdiffindiff_crossfit_inside(
  y0sample1,
  y1sample1,
  treatsample1,
  xsample1,
  zsample1,
  y0sample2,
  y1sample2,
  treatsample2,
  xsample2,
  zsample2,
  method,
  q
)
}
\arguments{
\item{y0sample1}{outcome at time 0 from sample 1}

\item{y1sample1}{outcome at time 1 from sample 1}

\item{treatsample1}{treatment status at time 1 from sample 1}

\item{xsample1}{pretreatment high dimensional covariates from sample 1}

\item{zsample1}{pretreatment low diemsioanl covariates allowing nonparametric specification from sample 1}

\item{y0sample2}{outcome at time 0 from sample 2}

\item{y1sample2}{outcome at time 1 from sample 2}

\item{treatsample2}{treatment status at time 1 from sample 2}

\item{xsample2}{pretreatment high dimensional covariates from sample 2}

\item{zsample2}{pretreatment low diemsioanl covariates allowing nonparametric specification from sample 2}

\item{method}{basis function in nonparametric specification:
\itemize{
\item (default) Pol: Polynomial basis with parameter q/2 as the degree of polynomial;
\item Tri: Trigonometric polynomials with parameter q/2 as the degree of polynomial;
}}

\item{q}{paramter for the basis function in nonparametric specification}
}
\value{
\item{xdebias}{debiased estimator for the high dimensional coefficients}
\item{gdebias}{debiased estimator for the nonparametric coefficients}
\item{stdx}{standard error for the high dimensional coefficients}
\item{stdg}{standard error for the nonparametric coefficients}
}
\description{
The core function of doubly robust diff-in-diff estimator under high dimesional covariates
}
\author{
Sida Peng
}
